ペンアイコンを削除．「続きを書く」ボタンに
