+++ cspell.json
@@ -169,6 +169,7 @@
     "recorded",
     "reduced",
     "registered",
+    "Relations",
     "remaining",
     "remains",
     "Remy",
@@ -197,6 +198,7 @@
     "Story",
     "styled",
     "stylelint",
+    "Submitted",
     "successfully",
     "Switch",
     "Syntheric",
