+++ src/features/NovelView/NovelViewPresentation.tsx
@@ -1,9 +1,10 @@
 import React, { useState } from 'react';
-import { Box, TextField, Button, Modal, Typography } from '@mui/material';
+import { Box, Button } from '@mui/material';
 import ParentPanel from './ParentPanel';
 import MainPanel from './MainPanel';
 import ChildrenPanel from './ChildrenPanel';
-import { NovelProps, Sentence } from '../../types/types';
+import { NovelProps, Sentence, CreateSentenceRequest } from '../../types/types';
+import { EditPost } from '../EditPost';
 
 interface NovelViewPresentationProps {
   mainPanel: Sentence[];
@@ -69,27 +70,16 @@ const NovelViewPresentation: React.FC<NovelViewPresentationProps> = ({
   onPost,
 }) => {
   const [start_index_main, setStart_index_main] = useState(0);
-  const [newSentence, setNewSentence] = useState('');
   const [isModalOpen, setIsModalOpen] = useState(false);
+  const [mainPanelState, setMainPanelState] = useState<Sentence[]>(mainPanel);
+  const [childrenPanelState, setChildrenPanelState] =
+    useState<Sentence[]>(childrenPanel);
 
-  const handlePost = async () => {
-    if (newSentence.trim()) {
-      await onPost(newSentence);
-      setNewSentence('');
+  const handlePost = async (sentenceRequest: CreateSentenceRequest) => {
+    if (sentenceRequest.text.trim()) {
+      await onPost(sentenceRequest.text);
       setIsModalOpen(false);
     }
-  };
-
-  const modalStyle = {
-    position: 'absolute' as 'absolute',
-    top: '50%',
-    left: '50%',
-    transform: 'translate(-50%, -50%)',
-    width: 400,
-    bgcolor: 'background.paper',
-    boxShadow: 24,
-    p: 4,
-    borderRadius: 2,
   };
 
   return (
@@ -161,49 +151,17 @@ const NovelViewPresentation: React.FC<NovelViewPresentationProps> = ({
             >
               続きを書く
             </Button>
-            <Modal
+            <EditPost
               open={isModalOpen}
               onClose={() => setIsModalOpen(false)}
-              aria-labelledby='novel-post-modal'
-            >
-              <Box sx={modalStyle}>
-                <Typography variant='h6' component='h2' gutterBottom>
-                  続きを書く
-                </Typography>
-                <TextField
-                  fullWidth
-                  multiline
-                  rows={6}
-                  value={newSentence}
-                  onChange={(e) => setNewSentence(e.target.value)}
-                  placeholder='新しい文章を入力してください'
-                  variant='outlined'
-                  sx={{ mb: 2 }}
-                />
-                <Box
-                  sx={{ display: 'flex', justifyContent: 'flex-end', gap: 1 }}
-                >
-                  <Button
-                    variant='outlined'
-                    onClick={() => setIsModalOpen(false)}
-                  >
-                    キャンセル
-                  </Button>
-                  <Button
-                    variant='contained'
-                    color='primary'
-                    onClick={handlePost}
-                  >
-                    投稿する
-                  </Button>
-                </Box>
-              </Box>
-            </Modal>
+              onSubmit={handlePost}
+              mainText={mainPanel[0]?.sentence || ''}
+            />
             <ChildrenPanel
-              childrenPanel={childrenPanel}
-              setChildrenPanel={() => {}}
-              mainPanel={mainPanel}
-              setMainPanel={() => {}}
+              childrenPanel={childrenPanelState}
+              setChildrenPanel={setChildrenPanelState}
+              mainPanel={mainPanelState}
+              setMainPanel={setMainPanelState}
               startIndex={start_index_children}
               setStartIndex={setStart_index_children}
               visibleTextCount={3}
